refact: Separate postprocessed functions
