+++ src/@types/params.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express'
 import { Dictionary } from 'lodash'
-import { UserNoPw } from './models'
+import { UserNoPw, User } from './models'
 
 export interface ReqAuth extends Request<Dictionary<string>> {
   body: UserNoPw
@@ -35,7 +35,13 @@ export type ListForm = {
   outside_seq: 0 | 1
 }
 
+export type ListFormOmit = Omit<ListForm, 'mode'>
+
 export type SearchOrList = SearchForm | ListForm
+
+export type PostList = Omit<User, 'semesters' | 'mailid'> & {
+  semestersReqParams: ListFormOmit[]
+}
 
 export interface FechParam extends UserNoPw {
   form: SearchOrList
